requirements.txt
aiohttp
anytree
aria2p
attrdict
appdirs
bs4
beautifulsoup4
cloudscrape
feedparser
gitpython
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
heroku3
hurry.filesize
itertools
js2py
lk21
jikanpy
lxml
pillow
psutil
psycopg2-binary
pybase64
pyrogram
pyshorteners
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
speedtest-cli
telegraph
tenacity
TgCrypto
torrentool==1.1.0
urllib3
youtube_dl
